Workaround &shellslash option on Windows (issue #6)
I'm not sure if this is the right thing to do but I've had several
"bug reports" about this option breaking my plug-ins. What I'm not
sure about is whether this option has a legitimate use case that
I'm breaking with this change...

TODO If I don't get any complaints about this change I may have to
move xolox#shell#escape() to xolox#misc#os#shellescape() and use it
in my other plug-ins as well?